More settings files in settings folder, fix cyclic import issue
