+++ apps/desktop-wallet/src/api/apiDataHooks/address/useFetchAddressLatestTransaction.ts
@@ -21,7 +21,7 @@ import { useQuery } from '@tanstack/react-query'
 import { UseFetchAddressProps } from '@/api/apiDataHooks/address/addressApiDataHooksTypes'
 import { addressLatestTransactionQuery } from '@/api/queries/transactionQueries'
 import { useAppSelector } from '@/hooks/redux'
-import { selectCurrentlyOnlineNetworkId } from '@/storage/settings/networkSelectors'
+import { selectCurrentlyOnlineNetworkId } from '@/storage/network/networkSelectors'
 
 const useFetchAddressLatestTransaction = ({ addressHash, skip }: UseFetchAddressProps) => {
   const networkId = useAppSelector(selectCurrentlyOnlineNetworkId)
