+++ apps/desktop-wallet/src/features/dataPolling/useAddressesDataPolling.ts
@@ -22,7 +22,7 @@ import { useQueries } from '@tanstack/react-query'
 import { addressLatestTransactionQuery } from '@/api/queries/transactionQueries'
 import { useAppSelector } from '@/hooks/redux'
 import { useUnsortedAddressesHashes } from '@/hooks/useAddresses'
-import { selectCurrentlyOnlineNetworkId } from '@/storage/settings/networkSelectors'
+import { selectCurrentlyOnlineNetworkId } from '@/storage/network/networkSelectors'
 
 const useAddressesDataPolling = () => {
   const allAddressHashes = useUnsortedAddressesHashes()
