+++ apps/desktop-wallet/src/features/dataPolling/usePendingTxPolling.ts
@@ -26,7 +26,7 @@ import { sentTransactionStatusChanged } from '@/features/send/sentTransactions/s
 import { selectSentTransactionByHash } from '@/features/send/sentTransactions/sentTransactionsSelectors'
 import { useAppDispatch, useAppSelector } from '@/hooks/redux'
 import { useUnsortedAddressesHashes } from '@/hooks/useAddresses'
-import { selectCurrentlyOnlineNetworkId } from '@/storage/settings/networkSelectors'
+import { selectCurrentlyOnlineNetworkId } from '@/storage/network/networkSelectors'
 
 const usePendingTxPolling = (txHash: Transaction['hash']) => {
   const dispatch = useAppDispatch()
