+++ apps/desktop-wallet/src/modals/SettingsModal/DevToolsSettingsSection.tsx
@@ -36,7 +36,7 @@ import { openModal } from '@/features/modals/modalActions'
 import { useAppDispatch, useAppSelector } from '@/hooks/redux'
 import { selectAllAddresses, selectDefaultAddress } from '@/storage/addresses/addressesSelectors'
 import { copiedToClipboard, copyToClipboardFailed, receiveTestnetTokens } from '@/storage/global/globalActions'
-import { devToolsToggled } from '@/storage/settings/settingsActions'
+import { devToolsToggled } from '@/features/settings/settingsActions'
 import { Address } from '@/types/addresses'
 
 const DevToolsSettingsSection = () => {
