+++ apps/desktop-wallet/src/storage/network/networkSelectors.ts
