+++ apps/desktop-wallet/src/storage/store.ts
@@ -27,8 +27,8 @@ import addressesSlice from '@/storage/addresses/addressesSlice'
 import contactsSlice from '@/storage/addresses/contactsSlice'
 import globalSlice from '@/storage/global/globalSlice'
 import snackbarSlice from '@/storage/global/snackbarSlice'
-import { networkListenerMiddleware } from '@/storage/settings/networkMiddleware'
-import settingsSlice, { settingsListenerMiddleware } from '@/storage/settings/settingsSlice'
+import { networkListenerMiddleware } from '@/storage/network/networkMiddleware'
+import settingsSlice, { settingsListenerMiddleware } from '@/features/settings/settingsSlice'
 import activeWalletSlice from '@/storage/wallets/activeWalletSlice'
 
 // TODO: Remove deconstruction when API related slices are removed from the shared reducer
