+++ apps/desktop-wallet/src/utils/migration.ts
@@ -28,12 +28,14 @@ import { AddressMetadata, Contact, NetworkSettings, networkSettingsPresets } fro
 import { merge } from 'lodash'
 import { nanoid } from 'nanoid'
 
+import { defaultSettings } from '@/features/settings/settingsConstants'
+import SettingsStorage from '@/features/settings/settingsPersistentStorage'
+import { GeneralSettings } from '@/features/settings/settingsTypes'
 import { addressMetadataStorage } from '@/storage/addresses/addressMetadataPersistentStorage'
 import { contactsStorage } from '@/storage/addresses/contactsPersistentStorage'
-import SettingsStorage, { defaultSettings } from '@/storage/settings/settingsPersistentStorage'
 import { walletStorage } from '@/storage/wallets/walletPersistentStorage'
 import { DeprecatedAddressMetadata } from '@/types/addresses'
-import { GeneralSettings, ThemeSettings } from '@/types/settings'
+import { ThemeSettings } from '@/types/theme'
 import { StoredEncryptedWallet } from '@/types/wallet'
 import { getRandomLabelColor } from '@/utils/colors'
 import { stringToDoubleSHA256HexString } from '@/utils/misc'
